backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langgraph
langchain-text-splitters
qdrant-client==1.12.0
redis
falkordb
flashrank
python-multipart
requests
numpy
tiktoken
python-dotenv
sqlalchemy
psycopg2-binary
pypdf
python-docx
pandas
openpyxl
oci
